refactor: コンポーネント名をケバブケースに変更
