+++ client/src/features/onboarding/pages/top/components/auth-pane/auth-pane.css.ts
