+++ client/src/features/onboarding/pages/top/components/auth-pane/index.ts
@@ -1,0 +1,1 @@
+export * from "./auth-pane";
