--- client/src/features/onboarding/pages/top/components/authPane/index.ts
@@ -1,1 +1,0 @@
-export * from "./authPane";
