+++ client/src/features/onboarding/pages/top/components/feature-pane/feature-pane.css.ts
