+++ client/src/features/onboarding/pages/top/components/feature-pane/feature-pane.tsx
@@ -1,5 +1,5 @@
 import { MaximumIcon } from "@/components/icons/maximum";
-import { styles } from "./featurePane.css";
+import { styles } from "./feature-pane.css";
 
 export const FeaturePane = () => {
   return (
