+++ client/src/features/onboarding/pages/top/components/feature-pane/index.ts
@@ -1,0 +1,1 @@
+export * from "./feature-pane";
