--- client/src/features/onboarding/pages/top/components/featurePane/index.ts
@@ -1,1 +1,0 @@
-export * from "./featurePane";
